Formal rules for splitting and merging SUMI packets
As discussed in yesterday's meeting, here is a proposal for the formal rules for splitting and merging SUMI packets.  My intent with this PR is just to clarify the spec, rather than change its meaning.  As a result, I left the meaning of SA as-is - let's consider the changes we discussed related to that in a separate PR.