Remove SA from responses
This PR proposes that SA be removed from UMI responses.  This provides an optimization opportunity for CLINK (not having to transmit SA for responses) and may simplify the implementation of RTL modules that manipulate SUMI packets.  At the SUMI level, the SA bus is always physically present, but implementations would be prohibited from depending on its value in responses.